+++ src/alloc.rs
@@ -1,5 +1,5 @@
 pub mod heap;
-use crate::core::def::NSTDAny;
+use crate::core::def::{NSTDAny, NSTDErrorCode};
 
 /// Allocates a new memory block.
 /// Parameters:
@@ -44,7 +44,7 @@ pub unsafe extern "C" fn nstd_alloc_allocate_zeroed(size: usize) -> NSTDAny {
 ///     `NSTDAny *const ptr` - Pointer to the memory block.
 ///     `const NSTDUSize size` - The current size of the memory block.
 ///     `const NSTDUSize new_size` - The new size of the memory block.
-/// Returns: `NSTDInt32 errc` - Nonzero on error.
+/// Returns: `NSTDErrorCode errc` - Nonzero on error.
 #[inline]
 #[cfg_attr(feature = "clib", no_mangle)]
 #[cfg_attr(target_os = "windows", allow(unused_variables))]
@@ -52,7 +52,7 @@ pub unsafe extern "C" fn nstd_alloc_reallocate(
     ptr: &mut NSTDAny,
     size: usize,
     new_size: usize,
-) -> i32 {
+) -> NSTDErrorCode {
     #[cfg(not(target_os = "windows"))]
     {
         use std::alloc::Layout;
@@ -76,11 +76,11 @@ pub unsafe extern "C" fn nstd_alloc_reallocate(
 /// Parameters:
 ///     `NSTDAny *const ptr` - Pointer to the memory block.
 ///     `const NSTDUSize size` - Number of bytes to deallocate.
-/// Returns: `NSTDInt32 errc` - Nonzero on error.
+/// Returns: `NSTDErrorCode errc` - Nonzero on error.
 #[inline]
 #[cfg_attr(feature = "clib", no_mangle)]
 #[cfg_attr(target_os = "windows", allow(unused_variables))]
-pub unsafe extern "C" fn nstd_alloc_deallocate(ptr: &mut NSTDAny, size: usize) -> i32 {
+pub unsafe extern "C" fn nstd_alloc_deallocate(ptr: &mut NSTDAny, size: usize) -> NSTDErrorCode {
     #[cfg(not(target_os = "windows"))]
     {
         use std::alloc::Layout;